// src/generated/client/types/ExerciseHistoryEntry.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type ExerciseHistoryEntry = {
    /**
     * @description The workout ID
     * @type string | undefined
    */
    workout_id?: string;
    /**
     * @description The workout title
     * @type string | undefined
    */
    workout_title?: string;
    /**
     * @description ISO 8601 timestamp of when the workout was recorded to have started.
     * @type string | undefined
    */
    workout_start_time?: string;
    /**
     * @description ISO 8601 timestamp of when the workout was recorded to have ended.
     * @type string | undefined
    */
    workout_end_time?: string;
    /**
     * @description The exercise template ID
     * @type string | undefined
    */
    exercise_template_id?: string;
    /**
     * @description The weight in kilograms
     * @type number
    */
    weight_kg?: number | null;
    /**
     * @description The number of repetitions
     * @type integer
    */
    reps?: number | null;
    /**
     * @description The distance in meters
     * @type integer
    */
    distance_meters?: number | null;
    /**
     * @description The duration in seconds
     * @type integer
    */
    duration_seconds?: number | null;
    /**
     * @description The Rating of Perceived Exertion
     * @type number
    */
    rpe?: number | null;
    /**
     * @description A custom metric for the set
     * @type number
    */
    custom_metric?: number | null;
    /**
     * @description The type of set (warmup, normal, failure, dropset)
     * @type string | undefined
    */
    set_type?: string;
};