// src/generated/client/types/ExerciseTemplate.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type ExerciseTemplate = {
    /**
     * @description The exercise template ID.
     * @type string | undefined
    */
    id?: string;
    /**
     * @description The exercise title.
     * @type string | undefined
    */
    title?: string;
    /**
     * @description The exercise type.
     * @type string | undefined
    */
    type?: string;
    /**
     * @description The primary muscle group of the exercise.
     * @type string | undefined
    */
    primary_muscle_group?: string;
    /**
     * @description The secondary muscle groups of the exercise.
     * @type array | undefined
    */
    secondary_muscle_groups?: string[];
    /**
     * @description A boolean indicating whether the exercise is a custom exercise.
     * @type boolean | undefined
    */
    is_custom?: boolean;
};